requirements.txt
yfinance
pandas
pyyaml
streamlit
plotly
altair<5